Move src/renderer/stores to src/stores
Cannoicalize the path of MobX stores.
