--- src/v2/stores
@@ -1,1 +1,0 @@
-../renderer/stores
