+++ src/v2/utils/useStore.tsx
@@ -1,7 +1,7 @@
 import React, { createContext, useContext } from "react";
-import AccountStore from "../stores/account";
-import GameStore from "../stores/game";
-import StandaloneStore from "../stores/standaloneStore";
+import AccountStore from "src/stores/account";
+import GameStore from "src/stores/game";
+import StandaloneStore from "src/stores/standaloneStore";
 
 const stores = {
   account: new AccountStore(),
